+++ src/map/vending.c
@@ -233,7 +233,7 @@ void vending_openvending(struct map_session_data *sd,int len,char *message,int f
 			}
 			sd->vending[i].amount = *(short*)(p+2+8*j);
 			sd->vending[i].value = *(int*)(p+4+8*j);
-			if(sd->vending[i].value > battle_config.vending_max_value)
+			if(sd->vending[i].value > (int)battle_config.vending_max_value)
 				sd->vending[i].value=battle_config.vending_max_value;
 			else if(sd->vending[i].value < 1)
 				sd->vending[i].value = 1000000;	// auto set to 1 million [celest]
